Hindrer client-side import av NextImageBuildTime
